A matrix meeting its own inverse now cancels
The legacy ACES reference fixture caught a fidelity fault no hand-built row
had. A config declares one direction of a colour space and Lumit derives the
other by inverting it, so `ACES - ACEScc -> ACES - ACEScg` resolves to the
AP1->AP0 matrix immediately followed by its own inverse: mathematically
nothing at all. Evaluated one step after the other in single precision it is
not nothing. At the probe [2.0, 1.4, 0.3] red has clamped to the top of the
ACEScc table at 65504, and blue detours through an intermediate of -246
before coming back to 0.045, carrying 2.3e-5 of that intermediate's rounding
with it: five hundred times the row's own tolerance.

Two halves to the fix, and both are needed. Matrix coefficients are held in
double, so an inverse is exact rather than rounded to single before anyone
composes with it; and Chain::new folds neighbouring matrices into one, so the
pair is composed rather than walked. Evaluation is unchanged: apply rounds
the twelve coefficients to single first, which is exactly what the graphics
card multiplies, so preview still equals export (K-031). The reference
library composes adjacent matrices for the same reason, so this is agreement
with it rather than a Lumit invention.

The spi1d reader was never at fault. Its interpolation reproduces the
reference to the last printed digit.

+++ crates/lumit-colour/src/clf.rs
@@ -179,10 +179,11 @@ fn matrix_op(what: &str, node: &Node) -> Result<Op> {
     let scale = in_scale / out_scale;
     for row in 0..3 {
         for col in 0..3 {
-            m[row * 4 + col] = node.array.get(row * cols + col).copied().unwrap_or(0.0) * scale;
+            m[row * 4 + col] =
+                f64::from(node.array.get(row * cols + col).copied().unwrap_or(0.0) * scale);
         }
         m[row * 4 + 3] = if cols == 4 {
-            node.array.get(row * cols + 3).copied().unwrap_or(0.0) / out_scale
+            f64::from(node.array.get(row * cols + 3).copied().unwrap_or(0.0) / out_scale)
         } else {
             0.0
         };
